20250806_RhCali_Shipment_Changes getDeclaredValue

+++ Shipment.java
@@ -35,7 +35,7 @@ public class Shipment{
         return cargoDescription;
     }
 
-    public long getDeclaredValue(){
+    public double getDeclaredValue(){
         return declaredValue;
     }
